to test deploy

+++ frontend/src/app/components/login-form/login-form.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router } from '@angular/router';
 import { LoginService } from 'src/app/services/login.service';
 import { User } from '../../clases/user';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -10,9 +11,9 @@ import { User } from '../../clases/user';
   styleUrls: ['./login-form.component.scss'],
 })
 export class LoginFormComponent implements OnInit {
-  public user=new User();
-  public contenido:string="Entrar";
-  public loginIcon="../../../assets/icon/log-in.png"; 
+  public user = new User();
+  public contenido: string = "Entrar";
+  public loginIcon = "../../../assets/icon/log-in.png";
 
   constructor(private loginService: LoginService,
     private router: Router) { }
@@ -20,12 +21,44 @@ export class LoginFormComponent implements OnInit {
   ngOnInit() { }
   public submit(): void {
     this.loginService.login(this.user).subscribe(
-      (data: number) => {
-        localStorage.setItem('userName', this.user.name);
-        localStorage.setItem('personalToken', `${data}`);
-        this.router.navigate(['/products']);
+      (data: String) => {
+        let timerInterval;
+        Swal.fire({
+          title: 'Iniciando Sesión!',
+          timer: 1000,
+          timerProgressBar: true,
+          didOpen: () => {
+            Swal.showLoading()
 
-
+          }
+        }).then(() => {
+          console.log(data);
+          if (data != null) {
+            localStorage.setItem('userName', this.user.email);
+            localStorage.setItem('personalToken', `${data}`);
+            this.router.navigate(['/products']);
+            Swal.fire({
+              title: 'Sesion iniciada con exito!',
+              icon: 'success',
+              timer:1000
+            })
+              .then(() => {
+                window.location.reload()
+              })
+          } 
+          else 
+          {
+            Swal.fire({
+              title: 'Error al iniciar sesión!',
+              text: 'El usuario o contraseña no existen',
+              timer: 5000,
+              icon: 'error',
+            })
+              .then(() => {
+                this.router.navigate(["/login"]);
+              })
+          }
+        })
       },
       () => {
         console.error("Error al realizar el acceso- login");
